requirements.txt
telethon==1.29.2
python-telegram-bot==20.3
nest_asyncio==1.5.6
